requirements.txt
APScheduler==3.9.1
astroid==2.11.1
attrs==21.4.0
autopep8==1.6.0
cachetools==5.0.0
certifi==2021.10.8
colorama==0.4.4
commonmark==0.9.1
configobj==5.0.6
dill==0.3.4
dnspython==2.2.1
flake8==4.0.1
iniconfig==1.1.1
isort==5.10.1
jedi==0.18.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
npyscreen==4.10.5
packaging==21.3
parso==0.8.3
pip-review==1.1.1
platformdirs==2.5.1
pluggy==1.0.0
PTable==0.9.2
py==1.11.0
pycodestyle==2.8.0
pydocstyle==6.1.1
pyflakes==2.4.0
Pygments==2.11.2
pylint==2.13.0
pymongo==4.0.2
pyparsing==3.0.7
pytest==7.1.1
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.4.0
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
rich==12.0.1
rope==0.23.0
six==1.16.0
snowballstemmer==2.2.0
telegram==0.0.1
terminaltables==3.1.10
toml==0.10.2
tomli==2.0.1
tornado==6.1
tzdata==2022.1
tzlocal==4.1
ujson==5.1.0
wcwidth==0.2.5
wrapt==1.14.0
yapf==0.32.0
